+++ apps/router-e2e/__e2e__/link-preview/app/(tabs)/home/slot/index.tsx
@@ -2,8 +2,6 @@ import { usePathname } from 'expo-router';
 import React from 'react';
 import { View, Text } from 'react-native';
 
-import { Links } from '@/__e2e__/link-preview/components/Links';
-
 export default function Index() {
   const pathname = usePathname();
 
@@ -11,7 +9,6 @@ export default function Index() {
     <View style={{ flex: 1, padding: 10, gap: 10, backgroundColor: '#dff' }}>
       <Text>Index</Text>
       <Text>Current Path: {pathname}</Text>
-      <Links />
     </View>
   );
 }