+++ apps/router-e2e/__e2e__/link-preview/app/(tabs)/settings/index.tsx
@@ -2,16 +2,13 @@ import { usePathname } from 'expo-router';
 import React from 'react';
 import { View, Text } from 'react-native';
 
-import { Links } from '@/__e2e__/link-preview/components/Links';
-
 const SettingsIndex = () => {
   const pathname = usePathname();
 
   return (
-    <View>
+    <View style={{ flex: 1, backgroundColor: '#fff' }}>
       <Text>Settings - Index</Text>
       <Text>Current Path: {pathname}</Text>
-      <Links />
     </View>
   );
 };