+++ packages/expo-router/ios/PeekAndPop/PeekAndPopTriggerView.swift
@@ -1,9 +1,0 @@
-import ExpoModulesCore
-import WebKit
-
-class PeekAndPopTriggerView: ExpoView {
-  required init(appContext: AppContext? = nil) {
-    super.init(appContext: appContext)
-    clipsToBounds = true
-  }
-}